+++ include/GafferSceneUI/TransformTool.h
@@ -40,6 +40,7 @@
 #include "GafferSceneUI/SelectionTool.h"
 #include "GafferSceneUI/TypeIds.h"
 
+#include "GafferScene/SceneAlgo.h"
 #include "GafferScene/ScenePlug.h"
 
 #include "GafferUI/KeyEvent.h"
@@ -88,52 +89,80 @@ class GAFFERSCENEUI_API TransformTool : public GafferSceneUI::SelectionTool
 			/// ============
 			///
 			/// The scene being viewed.
-			GafferScene::ConstScenePlugPtr scene;
-			/// The location within the viewed scene that has been
-			/// selected for editing.
-			GafferScene::ScenePlug::ScenePath path;
+			const GafferScene::ScenePlug *scene() const;
+			/// The location within the viewed scene that is being
+			/// edited.
+			const GafferScene::ScenePlug::ScenePath &path() const;
 			/// The context the scene is being viewed in.
-			Gaffer::ConstContextPtr context;
+			const Gaffer::Context *context() const;
 
 			/// Upstream scene
 			/// ==============
 			///
 			/// Often, the scene being viewed isn't actually the
-			/// scene that is being edited. Instead, an upstream
-			/// node is being edited, and the user is viewing a
-			/// downstream node to see the edits in the context of later
-			/// changes. The `upstreamScene` is the output from the node
-			/// actually being edited.
-			GafferScene::ConstScenePlugPtr upstreamScene;
+			/// scene where the transform originates. Instead, the
+			/// transform originates with an upstream node, and the
+			/// user is viewing a downstream node to see the transform
+			/// in the context of later changes. The `upstreamScene`
+			/// is the output from the node where the transform
+			/// originates.
+			const GafferScene::ScenePlug *upstreamScene() const;
 			/// The hierarchies of the upstream and viewed scenes may
 			/// differ. The upstreamPath is the equivalent of
 			/// the viewed path but in the upstream scene.
-			GafferScene::ScenePlug::ScenePath upstreamPath;
+			const GafferScene::ScenePlug::ScenePath &upstreamPath() const;
 			/// The upstream context is the equivalent of the
 			/// viewed context, but for the upstream scene.
-			Gaffer::ConstContextPtr upstreamContext;
+			const Gaffer::Context *upstreamContext() const;
 
-			/// Transform to edit
-			/// =================
-			///
-			/// The plug to edit. This will be a child of
-			/// the node generating the upstream scene.
-			Gaffer::TransformPlugPtr transformPlug;
-			/// The coordinate system within which the
+			/// Status and editing
+			/// ==================
+
+			/// Returns true if the selected transform may be edited
+			/// using `transformPlug()` and `transformSpace()`.
+			bool editable() const;
+			/// Returns a warning message, or "" if there are no
+			/// warnings.
+			const std::string &warning() const;
+
+			/// Returns the plug to edit. Throws if `!editable()`.
+			Gaffer::TransformPlug *transformPlug() const;
+			/// Returns the coordinate system within which the
 			/// transform is applied by the upstream node.
 			/// This is relative to the world space of the
-			/// upstream scene.
-			Imath::M44f transformSpace;
+			/// upstream scene. Throws if `!editable()`.
+			const Imath::M44f &transformSpace() const;
 
 			/// Utilities
 			/// =========
 			///
 			/// Returns a matrix which converts from world
 			/// space in `scene` to `transformSpace`.
+			/// Throws if `!editable()`.
 			Imath::M44f sceneToTransformSpace() const;
 			/// Returns a matrix suitable for positioning
 			/// transform handles in `scene's` world space.
+			/// Throws if `!editable()`.
 			Imath::M44f orientedTransform( Orientation orientation ) const;
+
+			private :
+
+				bool init( const GafferScene::SceneAlgo::History *history );
+				bool initWalk( const GafferScene::SceneAlgo::History *history );
+				void throwIfNotEditable() const;
+
+				GafferScene::ConstScenePlugPtr m_scene;
+				GafferScene::ScenePlug::ScenePath m_path;
+				Gaffer::ConstContextPtr m_context;
+
+				GafferScene::ConstScenePlugPtr m_upstreamScene;
+				GafferScene::ScenePlug::ScenePath m_upstreamPath;
+				Gaffer::ConstContextPtr m_upstreamContext;
+
+				bool m_editable;
+				std::string m_warning;
+				Gaffer::TransformPlugPtr m_transformPlug;
+				Imath::M44f m_transformSpace;
 
 		};