+++ python/GafferSceneUITest/__init__.py
@@ -49,6 +49,7 @@ from RotateToolTest import RotateToolTest
 from ContextAlgoTest import ContextAlgoTest
 from CameraToolTest import CameraToolTest
 from VisualiserTest import VisualiserTest
+from TransformToolTest import TransformToolTest
 
 if __name__ == "__main__":
 	unittest.main()