+++ Changes.md
@@ -28,11 +28,15 @@ API
 - ScenePlug : Added `exists()` method. This provides fast existence queries for locations.
 - SceneAlgo : Deprecated `exists()` function. Use `ScenePlug::exists()` instead
 - Spreadsheet : Added `RowsPlug::row( rowName )` method.
+- TransformTool :
+  - Added `Selection::editable()` method.
+  - Added `Selection::warning()` method.
 
 Breaking Changes
 ----------------
 
 - Context : Removed `Substitutions`, `substitutions()` and `hasSubstitutions()`. Use the `IECore.StringAlgo` equivalents instead.
+- TransformTool : The `Selection` class now uses accessor methods rather than exposing data members directly.
 - Stats app : The `-scene` and `-image` arguments now ignore input plugs.
 - Spreadsheet : Broke binary compatibility. Full source compatibility is retained.